+++ RUFAS/biophysical/feed_storage/feed_manager.py
@@ -233,6 +233,17 @@ class FeedManager:
         self.report_stored_farmgrown_feeds(simulation_day, reporting_suffix)
         self.purchased_feed_storage.report_stored_purchased_feeds(simulation_day, reporting_suffix)
 
+    def report_cumulative_purchased_feeds(self, simulation_day: int) -> None:
+        """Outputs the cumulative purchased feeds to the output manager."""
+        info_map = {
+            "class": self.__class__.__name__,
+            "function": self.report_cumulative_purchased_feeds.__name__,
+            "simulation_day": simulation_day,
+            "units": MeasurementUnits.KILOGRAMS,
+        }
+        for rufas_id, amount in self._cumulative_purchased_feeds.items():
+            self._om.add_variable(f"purchased_feed_{rufas_id}_purchased_to_date", amount, info_map)
+
     def report_stored_farmgrown_feeds(self, simulation_day: int, reporting_suffix: str) -> None:
         """Outputs total amounts of farmgrown feeds currently stored by the FeedManager."""
         feed_report: dict[RUFAS_ID, dict[str, float]] = {
@@ -502,11 +513,6 @@ class FeedManager:
 
             total_cost = purchase_amount * feed_info.purchase_cost
 
-            info_map = info_map | {
-                "price": feed_info.purchase_cost,
-                "amount_purchased": purchase_amount,
-                "total_cost": total_cost,
-            }
             self._om.add_variable(
                 f"{purchase_type}_{rufas_id}_cost",
                 total_cost,